+++ lib/runners/config.rb
@@ -35,6 +35,12 @@ module Runners
 
     FILE_NAME = "sider.yml".freeze
     FILE_NAME_OLD = "sideci.yml".freeze
+    DEFAULT_METRICS_IGNORE_PATTERNS = [
+      "**/node_modules/**",
+      "**/vendor/**",
+      "*.map",
+      "*.min.*",
+    ].freeze
 
     def self.load_from_dir(dir)
       file = [
@@ -56,6 +62,10 @@ module Runners
       @warning_checkers.each do |checker|
         checker.call(config)
       end
+    end
+
+    def self.invert_patterns(patterns)
+      patterns.map { |pat| pat.start_with?('!') ? pat.delete_prefix('!') : '!' + pat }
     end
 
     attr_reader :path, :raw_content
@@ -99,6 +109,10 @@ module Runners
 
     def ignore_patterns
       @ignore_patterns ||= Array(content[:ignore])
+    end
+
+    def metrics_ignore_patterns
+      @metrics_ignore_patterns ||= DEFAULT_METRICS_IGNORE_PATTERNS + Array(content.dig(:linter, :metrics, :ignore))
     end
 
     def linter(id)