+++ lib/runners/processor.rb
@@ -15,12 +15,16 @@ module Runners
         .freeze
     end
 
-    def self.register_config_schema(schema)
-      Schema::Config.register(name: analyzer_id, schema: schema)
+    def self.register_config_schema(schema, name: analyzer_id)
+      Schema::Config.register(name: name, schema: schema)
     end
 
     def self.config_example
       raise NotImplementedError, "#{name} needs to implement `.#{__method__}`"
+    end
+
+    def self.metrics?
+      [MetricsComplexity, MetricsCodeClone, MetricsFileInfo].include?(self)
     end
 
     attr_reader :guid, :working_dir, :config, :shell, :trace_writer, :warnings, :options