+++ test/smokes/metrics_complexity/with_ignore/assets/bar/hello.swift
@@ -1,0 +1,6 @@
+func hello_world() {
+/* Hello world in Swift */
+  print("Hello, World!")
+}
+
+hello_world()