+++ test/smokes/metrics_complexity/with_ignore/assets/hello.c
@@ -1,0 +1,17 @@
+#include <stdio.h>
+#include <stdlib.h>
+
+void print_hello_world();
+
+int main(void)
+{
+  print_hello_world();
+  return EXIT_SUCCESS;
+}
+
+void print_hello_world()
+{
+/* Hello world in C */
+  puts("Hello, World!");
+}
+