+++ test/smokes/metrics_complexity/with_ignore/example.rb
@@ -1,0 +1,6 @@
+def hello_world
+# Hello world in Ruby
+  puts "Hello, World!"
+end
+
+hello_world