+++ test/smokes/metrics_complexity/with_ignore/src/baz/qux/hello.lua
@@ -1,0 +1,7 @@
+
+function hello_world()
+-- Hello world in Lua
+  print "Hello, World!"
+end
+
+hello_world()