+++ test/smokes/metrics_complexity/with_ignore/src/baz/qux/hello.rs
@@ -1,0 +1,9 @@
+
+fn main() {
+  print_hello();
+}
+
+fn print_hello() {
+/* Hello world in Rust */
+  println!("Hello, World!");
+}