+++ test/smokes/metrics_fileinfo/expectations.rb
@@ -15,7 +15,7 @@ s.add_test_with_git_metadata(
         object: {
           lines_of_code: 7,
           last_committed_at: "2021-01-01T10:00:00+09:00",
-          number_of_commits: 0,
+          number_of_commits: 1,
           occurrence: 0,
           additions: 0,
           deletions: 0
@@ -322,6 +322,46 @@ s.add_test_with_git_metadata(
           occurrence: 0,
           additions: 0,
           deletions: 0
+        },
+        git_blame_info: nil
+      }
+    ],
+    analyzer: { name: "Metrics File Info", version: default_version } }
+)
+
+s.add_test_with_git_metadata(
+  "with_metrics_ignore_setting",
+  { type: "success",
+    issues: [
+      {
+        id: "metrics_fileinfo",
+        path: "hello.rb",
+        location: nil,
+        message: "hello.rb: loc = 7, last commit datetime = 2021-01-01T10:00:00+09:00",
+        links: [],
+        object: {
+          lines_of_code: 7,
+          last_committed_at: "2021-01-01T10:00:00+09:00",
+          number_of_commits: 1,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
+        },
+        git_blame_info: nil
+      },
+      {
+        id: "metrics_fileinfo",
+        path: "sider.yml",
+        location: nil,
+        message: "sider.yml: loc = 4, last commit datetime = 2021-08-03T02:19:42+00:00",
+        links: [],
+        object: {
+          lines_of_code: 4,
+          last_committed_at: "2021-08-03T02:19:42+00:00",
+          number_of_commits: 1,
+          occurrence: 1,
+          additions: 4,
+          deletions: 2
         },
         git_blame_info: nil
       }