+++ test/smokes/metrics_fileinfo/with_metrics_ignore_setting/hello.rb
@@ -1,0 +1,7 @@
+# A just simple ruby program
+def hello
+  puts "Hello, world"
+end
+
+hello
+