+++ test/smokes/metrics_fileinfo/with_metrics_ignore_setting/ignored_dir/ignored_hello.rb
@@ -1,0 +1,7 @@
+# A just simple ruby program
+def hello
+  puts "Hello, world, ignored"
+end
+
+hello
+